requirements.txt
requests
python-dotenv
fastapi
uvicorn
python-multipart
aiofiles
httpx
sqlalchemy
psycopg2-binary
mpmath